backend/requirements.txt
# M&A Research Pipeline - Backend Dependencies
# Core
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
beautifulsoup4==4.12.3

# Data processing
pandas==2.2.3

# Utilities
python-dotenv==1.0.1
